aegis-django/requirements/development.txt
# AEGIS Django - Development Requirements
-r base.txt

# Development tools
django-debug-toolbar==4.2.0
django-extensions==3.2.3
ipython==8.20.0

# Testing
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Celery monitoring
flower==2.0.1

# Code quality
black==24.1.1
flake8==7.0.0
isort==5.13.2
mypy==1.8.0
